feat(LLM): Expand bridge integration for M3-M5
C++ side of the LLM middleware integration milestones beyond M2:

- LlmBridgeHook: party / raid / guild chat call-sites added so the
  Python middleware wakes on every channel a bot is reachable on.
- PlayerbotAIConfig: AiPlayerbot.LlmBridgeEndpoint config key honored
  end-to-end (read at startup, used in the hook worker thread).
- PlayerbotAI / Value: party,guid TCP command returns
  Name:Class:HpPct:Mana for each party member so the Python combat
  context builder can render its [PARTY STATE] block without a
  separate request per member.
- LootRollAction: arbitration hook so the middleware's loot-roll
  arbitration can dispatch pass commands to losers and a single LLM
  call to the winner (M4).
- NewRpgAction: RPG mode hooks so set_rpg_mode / go_to_location tools
  on the Python side actually move the bot (M4).
- Chat action/trigger context headers: hooks for chat-derived events
  routed through the bridge.

+++ src/Bot/Engine/Value/Value.cpp
@@ -29,7 +29,11 @@ std::string const UnitManualSetValue::Format()
 std::string const Uint8CalculatedValue::Format()
 {
     std::ostringstream out;
-    out << Calculate();
+    // Cast to uint32 so the stream formats it as a decimal integer.
+    // Without the cast, uint8 (= unsigned char) would be emitted as a
+    // raw byte, so a count of 3 would appear as the ETX control char
+    // rather than the string "3".
+    out << static_cast<uint32>(Calculate());
     return out.str();
 }
